Add files via upload
The Word Counter is a powerful and user-friendly tool designed to count the number of words and characters in any text. This application is equipped with advanced features like Save, Clear, and the ability to manage privacy settings by making the project public or private. It is designed to be intuitive for users who need to analyze text quickly and efficiently, with flexible options for saving and managing their data.